requirements.txt
django-jquery
django-bootstrap-toolkit
django-taggit-autosuggest
django-mptt 
django-dag
south
# Uncomment if using MySQL
#MySQL-python
